Create toysamplefixedabs
This code demonstrates the successful creation of a toy corpus from a fixed National Library of Medicine efetch operation of three abstracts containing text with special characters that the program is able to parse, delimit, and transform successfully into a pre-processed list and string object using delimiters that work for non-structured abstracts and save to disk for corpus persistency. A simple demonstration of basic functions that were extensively built upon for subsequent large-scale formal text scraping of pharmaceutical text and other data.